Moved the classes to User module folder
