+++ src/main/java/br/com/caelum/carangobom/user/UserRepository.java
@@ -1,6 +1,6 @@
-package br.com.caelum.carangobom.user.repository;
+package br.com.caelum.carangobom.user;
 
-import br.com.caelum.carangobom.user.model.User;
+import br.com.caelum.carangobom.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface UserRepository extends JpaRepository<User, Long> {
